tests and fixes added

+++ jobworker/worker.go
@@ -17,19 +17,18 @@ type Worker struct {
 	ID         string
 	Cmd        *exec.Cmd
 	status     string
-	err        error
 	mu         sync.RWMutex
 	outputFile string
 }
 
 // Worker status string constants.
 const (
-	wPending  = "pending"
-	wStarted  = "started"
-	wRunning  = "running"
-	wKilled   = "killed"
-	wFinished = "finished"
-	wError    = "error"
+	WPending  = "pending"
+	WStarted  = "started"
+	WRunning  = "running"
+	WKilled   = "killed"
+	WFinished = "finished"
+	WError    = "error"
 )
 
 // NewWorker creates a new worker with an unique id and output file.
@@ -39,7 +38,13 @@ func NewWorker(command string, args ...string) (*Worker, error) {
 	var uid = uuid.NewString()
 
 	// Hardcoded a /logs directory to store worker output as temp files.
-	// Tempdir giving each worker a dir to log would also work.
+	// Making a tempdir and giving each worker a dir to log would also work.
+	err := mkdir("logs", os.FileMode(0777))
+
+	if err != nil {
+		return nil, err
+	}
+
 	of, err := ioutil.TempFile("logs", "worker_output_*")
 
 	if err != nil {
@@ -55,7 +60,7 @@ func NewWorker(command string, args ...string) (*Worker, error) {
 	return &Worker{
 		ID:         uid,
 		Cmd:        exec.Command(command, args...),
-		status:     wPending,
+		status:     WPending,
 		outputFile: of.Name(),
 	}, nil
 
@@ -64,15 +69,15 @@ func NewWorker(command string, args ...string) (*Worker, error) {
 // Start initiates a worker command and calls execute().
 func (w *Worker) Start() error {
 
-	w.UpdateStatus(wStarted)
+	w.UpdateStatus(WStarted)
 	err := w.execute()
 
 	if err != nil {
 		log.Print("Error: ", err)
-		w.UpdateStatus(wError)
+		w.UpdateStatus(WError)
 		return err
 	}
-	w.UpdateStatus(wFinished)
+	w.UpdateStatus(WFinished)
 
 	return nil
 }
@@ -90,35 +95,34 @@ func (w *Worker) execute() error {
 		return err
 	}
 
-	err = cmd.Start()
-	if err != nil {
-		return err
-	}
-
-	w.UpdateStatus(wRunning)
-
-	// copy combined process output to outputFile.
 	combinedOutput := io.MultiReader(stdout, stderr)
-
 	f, err := os.OpenFile(w.outputFile, os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
 
 	// Note: if error occurs when opening/closing/writing to the file, the worker would result in an error
 	// state if error was returned; However the process itself might still be running fine, and logging
-	// the errors instead would let the process proceed.
+	// the errors instead would let the process proceed in this case.
 	if err != nil {
 		log.Print(err)
 	}
 
-	// Note: same as above.
+	// Note: error handling same as above.
 	defer func() {
 		if err := f.Close(); err != nil {
 			log.Print(err)
 		}
 	}()
 
+	// start the command and update status to running.
+	err = cmd.Start()
+	if err != nil {
+		return err
+	}
+
+	w.UpdateStatus(WRunning)
+
 	_, err = io.Copy(f, combinedOutput)
 
-	// Note: same as above.
+	// Note: error handling same as above.
 	if err != nil {
 		log.Print(err)
 	}
@@ -137,32 +141,32 @@ func (w *Worker) execute() error {
 
 // Stop terminates a worker if it is running.
 func (w *Worker) Stop() error {
-	currentStatus, err := w.Status()
-	if err != nil {
-		return err
-	}
-	if currentStatus != wRunning {
+	var currentStatus = w.Status()
+	if currentStatus != WRunning {
 		return errors.New("Error: attempt to kill a process that is not running")
 	}
 
-	err = w.Cmd.Process.Signal(os.Kill)
+	if w.Cmd.Process == nil {
+		return errors.New("Error: unable to kill a nil process")
+	}
+	err := w.Cmd.Process.Signal(os.Kill)
 
 	if err != nil {
 		return err
 	}
 
-	w.UpdateStatus(wKilled)
+	w.UpdateStatus(WKilled)
 	return nil
 }
 
 // Status returns the worker's status, which is one of: pending, started, running, killed, error, finished.
-func (w *Worker) Status() (string, error) {
+func (w *Worker) Status() string {
 
 	// returns status with a lock.
 	w.mu.RLock()
 	defer w.mu.RUnlock()
 
-	return w.status, nil
+	return w.status
 
 }
 
@@ -171,7 +175,9 @@ func (w *Worker) UpdateStatus(status string) {
 	// update status with lock
 	w.mu.Lock()
 	defer w.mu.Unlock()
-	w.status = status
+	if w.status != WKilled {
+		w.status = status
+	}
 }
 
 // completeStatus returns the exit code of a worker with lock.
@@ -181,16 +187,35 @@ func (w *Worker) completeStatus() int {
 	return w.Cmd.ProcessState.ExitCode()
 }
 
-// RemoveLogs cleans up the output log with lock.
+// RemoveLogs cleans up the output log with lock. It is the caller's responsibility to remove them.
 func (w *Worker) RemoveLogs() error {
 	w.mu.RLock()
 	defer w.mu.RUnlock()
+	var currentStatus = w.Status()
+	if currentStatus == WRunning {
+		return errors.New("Error: attempting to remove the log of a running process")
+	}
 	return os.Remove(w.outputFile)
 }
 
-// ReadLogs returns a read closer of the output log file.
+// ReadLogs returns a read closer of the output log file with lock.
 func (w *Worker) ReadLogs() (io.ReadCloser, error) {
 	w.mu.RLock()
 	defer w.mu.RUnlock()
+	var currentStatus = w.Status()
+	if currentStatus == WRunning {
+		return nil, errors.New("Error: attempting to read the log of a running process")
+	}
 	return os.Open(w.outputFile)
+}
+
+// mkdir is used to make the directory if it does not exist
+func mkdir(name string, perm os.FileMode) error {
+	err := os.Mkdir(name, perm)
+	if err != nil {
+		if !os.IsExist(err) {
+			return err
+		}
+	}
+	return nil
 }
